requirements.txt
pyTelegramBotAPI>=4.14.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
thefuzz>=0.20.0
transformers>=4.35.0
torch>=2.0.0
faiss-cpu
re